requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dateutil>=2.8.0
requests>=2.25.0
beautifulsoup4>=4.9.0
